# api/routes/usuario.js
import express from 'express'
import { connectToDatabase } from '../utils/mongodb.js'
import { check, validationResult } from 'express-validator'
import bcrypt from 'bcryptjs'
import jwt from 'jsonwebtoken'
import auth from '../middleware/auth.js'

const router = express.Router()

const { db, ObjectId } = await connectToDatabase()
const nomeCollection = 'usuarios'

//############################################### VALIDAÇÕES DO USUÁRIO ##################################################

const validaUsuario = [
    check('nome')
        .not().isEmpty().trim().withMessage('É obrigatório informar o nome')
        .isAlpha('pt-BR', { ignore: ' ' }).withMessage('Informe apenas texto')
        .isLength({ min: 3 }).withMessage('Informe no mínimo 3 caracteres')
        .isLength({ max: 100 }).withMessage('Informe no máximo 100 caracteres'),
    check('email')
        .not().isEmpty().trim().withMessage('É obrigatório informar o email')
        .isLowercase().withMessage('Não são permitidas letras maiúsculas')
        .isEmail().withMessage('Informe um email válido')
        .custom((value, { req }) => {
            return db.collection(nomeCollection)
                .find({ email: { $eq: value } }).toArray()
                .then((email) => {
                    //verifica se não existe o ID para garantir que é inclusão
                    if (email.length && !req.params.id) {
                        return Promise.reject(`o email já existe!`)
                    }
                })
        }),
    check('senha')
        .not().isEmpty().trim().withMessage('A senha é obrigatória')
        .isLength({ min: 6 }).withMessage('A senha deve ter no mínimo 6 carac.')
        .isStrongPassword({
            minLength: 6,
            minLowercase: 1, minUppercase: 1,
            minSymbols: 1, minNumbers: 1
        }).withMessage('A senha não é segura. Informe no mínimo 1 caractere maiúsculo, 1 minúsculo, 1 número e 1 caractere especial'),
    check('ativo')
        .default(true)
        .isBoolean().withMessage('O valor deve ser um booleano'),
    check('tipo')
        .default('Cliente')
        .isIn(['Admin', 'Cliente']).withMessage('O tipo deve ser Admin ou Cliente'),
    check('avatar')
        .optional({ nullable: true })
        .isURL().withMessage('A URL do avatar é inválida')
]


//############################################### VALIDAÇÕES DO LOGIN ####################################################

const validaLogin = [
    check('email')
        .not().isEmpty().trim().withMessage('O email é obrigatório')
        .isEmail().withMessage('Informe um email válido para o login'),
    check('senha')
        .not().isEmpty().trim().withMessage('A senha é obrigatória')
]


//############################################### OPERAÇÕES DO POST ######################################################

/** 
 * POST /usuarios
 * Adiciona um novo usuário ao sistema
 * Parâmetros: Objeto usuario
 * Postman: http://localhost:4000/usuarios
*/

router.post('/', validaUsuario, async (req, res) => {
    /* 
    * #swagger.tags = ['Usuários']
    * #swagger.summary = 'Adiciona um novo usuário ao sistema'
    * #swagger.description = 'Endpoint para adicionar um novo usuário.'    
    * #swagger.path = '/usuarios'
    * #swagger.method = 'POST'
    */
    const schemaErrors = validationResult(req)
    if (!schemaErrors.isEmpty()) {
        return res.status(403).json({
            errors: schemaErrors.array()
        })
    } else {
        //definindo o avatar default
        req.body.avatar = `https://ui-avatars.com/api/?name=${req.body.nome.replace(/ /g, '+')}&background=F00&color=FFF`
        //criptografia da senha
        //genSalt => impede que 2 senhas iguais tenham resultados iguais
        const salt = await bcrypt.genSalt(10)
        req.body.senha = await bcrypt.hash(req.body.senha, salt)
        //iremos salvar o registro
        await db.collection(nomeCollection)
            .insertOne(req.body)
            .then(result => res.status(201).send(result))
            .catch(err => res.status(400).json(err))
    } //fecha o else
})


/**
 * POST /usuarios/login
 * Efetua o login do usuário do sistema
 * Parâmetros: Objeto JSON contendo "email" e "senha" do usuário.
 * Postman: http://localhost:4000/usuarios/login
 */

router.post('/login', validaLogin, async (req, res) => {
    /* 
    * #swagger.tags = ['Usuários']
    * #swagger.summary = 'Efetua o login do usuário do sistema'
    * #swagger.description = 'Endpoint para efetuar o login e receber o token JWT.'    
    * #swagger.path = '/usuarios/login'
    * #swagger.method = 'POST'
    */
    const schemaErrors = validationResult(req)
    if (!schemaErrors.isEmpty()) {
        return res.status(403).json(({ errors: schemaErrors.array() }))
    }
    //obtendo os dados para o login
    const { email, senha } = req.body
    try {
        //verificar se o email existe no Mongodb
        let usuario = await db.collection(nomeCollection)
            .find({ email }).limit(1).toArray()
        //Se o array estiver vazio, é que o email não existe
        if (!usuario.length)
            return res.status(404).json({ //not found
                errors: [{
                    value: `${email}`,
                    msg: `O email ${email} não está cadastrado!`,
                    param: 'email'
                }]
            })
        //Se o email existir, comparamos se a senha está correta
        const isMatch = await bcrypt.compare(senha, usuario[0].senha)
        if (!isMatch)
            return res.status(403).json({ //forbidden
                errors: [{
                    value: 'senha',
                    msg: 'A senha informada está incorreta',
                    param: 'senha'
                }]
            })
        //Iremos gerar o token JWT
        jwt.sign(
            { usuario: { id: usuario[0]._id } },
            process.env.SECRET_KEY,
            { expiresIn: process.env.EXPIRES_IN },
            (err, token) => {
                if (err) throw err
                res.status(200).json({
                    access_token: token
                })
            }
        )
    } catch (e) {
        console.error(e)
    }

})


/**
 * POST /logout
 * Efetua o logout do usuário
 * Parâmetros: Objeto JSON com mensagem do resultado da operação
 * Postman: http://localhost:4000/logout
 */

router.post('/logout', auth, async (req, res) => {
    /* 
    * #swagger.tags = ['Usuários']
    * #swagger.summary = 'Efetua o logout
    * #swagger.description = 'Endpoint para realizar logout'    
    * #swagger.path = '/usuarios'
    * #swagger.method = 'POST'
    */
    try {
        // Limpar o token JWT do cliente (por exemplo, do armazenamento local)
        localStorage.removeItem('jwtToken');

        // Opcional: invalidar o token no servidor (se estiver mantendo uma lista negra de tokens inválidos)
        // Adicione o token à lista negra ou faça qualquer outra ação necessária para invalidar o token no servidor

        res.status(200).json({ message: 'Logout bem-sucedido' });
    } catch (err) {
        res.status(500).json({ message: 'Erro ao fazer logout' });
    }
});


//############################################### OPERAÇÕES DO GET #######################################################

/**
 * GET /usuarios
 * Lista todos os usuários do sistema
 * Parâmetros: eq (objeto HTTP), res (objeto HTTP). Retorna: Objeto JSON com a lista de usuários ou mensagem de erro
 * Postman: http://localhost:4000/usuarios
 */

router.get('/', auth, async (req, res) => {
    /* 
    * #swagger.tags = ['Usuários']
    * #swagger.summary = 'Lista todos os usuários do sistema'
    * #swagger.description = 'Endpoint para obter todos os usuários do sistema.'    
    * #swagger.path = '/usuarios'
    * #swagger.method = 'GET'
    */
    try {
        const docs = []
        await db.collection(nomeCollection)
            .find({}, { senha: 0 })
            .sort({ nome: 1 })
            .forEach((doc) => {
                docs.push(doc)
            })
        res.status(200).json(docs)
    } catch (err) {
        res.status(500).json({
            message: 'Erro ao obter a listagem dos usuários',
            error: `${err.message}`
        })
    }
})


/**
 * GET /usuarios/id/:id
 * Exibe os detalhes do usuário logado
 * Postman: http://localhost:4000/api/usuarios/id/123 (substitua 123 pelo ID do usuário logado)
*/

router.get('/id/:id', auth, async (req, res) => {
    /* 
    * #swagger.tags = ['Usuários']
    * #swagger.summary = 'Exibe os detalhes do usuário logado'
    * #swagger.description = 'Endpoint para obter os detalhes do usuário logado.'    
    * #swagger.path = '/usuarios/id/{id}'
    * #swagger.method = 'GET'
    */
    try {
        // Busca o usuário pelo ID no banco de dados
        const query = { _id: ObjectId(req.params.id) };
        const user = await db.collection(nomeCollection).findOne(query, { projection: { senha: 0 } });

        // Verifica se o usuário existe
        if (!user) {
            return res.status(404).json({ msg: 'Usuário não encontrado' });
        }

        // Retorna os detalhes do usuário (sem a senha)
        res.status(200).json(user);
    } catch (err) {
        console.error(err); // Log do erro no console para debug
        res.status(500).json({
            errors: [{
                value: `${err.message}`,
                msg: 'Erro ao obter o usuário pelo ID',
                param: req.params.id
            }]
        });
    }
});




//############################################### OPERAÇÕES DO PUT #######################################################

/**
 * PUT /usuarios/
 * Permite edição dos dados do usuário
 * Parâmetros:
    * id (no path): ID do usuário a ser atualizado
    * Objeto JSON no corpo da requisição contendo os dados atualizados do usuário
 * Postman: http://localhost:4000/usuarios/123 (substitua 123 pelo ID do usuário desejado)
*/
router.put('/:id', auth, async (req, res) => {
    /* 
    * #swagger.tags = ['Usuários']
    * #swagger.summary = 'Permite edição dos dados do usuario'
    * #swagger.description = 'Endpoint para edita dados usuario.'    
    * #swagger.path = '/usuarios/id'
    * #swagger.method = 'PUT'
    */
    try {
        // Extrair o ID do parâmetro da URL
        const resourceId = req.params.id;

        // Extrair os dados a serem atualizados do corpo da requisição
        const updatedData = req.body;

        // Verificar se o ID fornecido é válido
        if (!ObjectId.isValid(resourceId)) {
            return res.status(400).json({ message: 'ID inválido' });
        }

        // Tentar atualizar o recurso no banco de dados
        const result = await db.collection(nomeCollection).updateOne(
            { _id: new ObjectId(resourceId) },
            { $set: updatedData }
        );

        // Verificar se o recurso foi encontrado e atualizado
        if (result.modifiedCount === 0) {
            return res.status(404).json({ message: 'Recurso não encontrado' });
        }

        // Responder com sucesso
        res.status(200).json({ message: 'Recurso atualizado com sucesso' });
    } catch (error) {
        console.error('Erro ao atualizar o recurso:', error);
        res.status(500).json({ message: 'Erro interno do servidor ao atualizar o recurso' });
    }
});


//############################################### OPERAÇÕES DO DELETE ####################################################

/**
 * DELETE /usuarios/
 * Exclui um usuário
 * Parâmetros: id (no path): ID do usuário a ser excluído
 * Postman: http://localhost:4000/usuarios/123 (substitua 123 pelo ID do usuário a ser excluído)
*/

router.delete('/:id', auth, async (req, res) => {
    /* 
    * #swagger.tags = ['Usuários']
    * #swagger.summary = 'Exclui um usuário'
    * #swagger.description = 'Endpoint para deletar usuario.'    
    * #swagger.path = '/usuarios/id'
    * #swagger.method = 'DELETE'
    */
    try {
      const result = await db.collection(nomeCollection).deleteOne({
        "_id": new ObjectId(req.params.id)
      });
  
      if (result.deletedCount === 0) {
        return res.status(404).json({
          errors: [{
            value: `Não há nenhum usuário com o id ${req.params.id}`,
            msg: 'Erro ao excluir o usuário',
            param: '/:id'
          }]
        });
      } else {
        return res.status(200).send(result);
      }
    } catch (err) {
      return res.status(500).json({
        errors: [{
          value: `${req.params.id}`,
          msg: 'Erro ao excluir o usuário',
          param: '/:id',
          location: 'params'
        }]
      });
    }
  });

  
// Exportar o módulo de rotas
export default router

  