# api/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "📊📈API CalmaMente",
    "description": "➡️Documentação gerada automaticamente pelo módulo <a href='https://github.com/davibaltar/swagger-autogen' target='_blank'>swagger-autogen</a>."
  },
  "host": "localhost:4000",
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "access-token",
      "description": "Token de Acesso gerado após o login"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Default"
        ],
        "summary": "Rota default que retorna a versão da API",
        "description": "Endpoint que retorna a versão da API",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/clinicas": {
      "post": {
        "tags": [
          "Clínicas"
        ],
        "summary": "Cria uma nova clínica",
        "description": "Endpoint para criar uma nova clínica.",
        "responses": {
          "201": {
            "description": "Criado"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      },
      "get": {
        "tags": [
          "Clínicas"
        ],
        "summary": "Obtém a lista de clínicas",
        "description": "Endpoint para obter a lista de todas as clínicas disponíveis.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "skip",
            "in": "query",
            "type": "string"
          },
          {
            "name": "order",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/clinicas/id/{id}": {
      "get": {
        "tags": [
          "Clínicas"
        ],
        "summary": "Obtém uma clínica pelo ID",
        "description": "Endpoint para obter os detalhes de uma clínica específica com base no ID fornecido.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/clinicas/nome/{filtro}": {
      "get": {
        "tags": [
          "Clínicas"
        ],
        "summary": "Obtém uma clínica pelo nome",
        "description": "Endpoint para obter os detalhes de uma clínica específica com base no nome fornecido.",
        "parameters": [
          {
            "name": "filtro",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/clinicas/especialidades/{especialidade}/uf/{uf}": {
      "get": {
        "tags": [
          "Clínicas"
        ],
        "summary": "Obtém clínicas por especialidade e UF",
        "description": "Endpoint para obter uma lista de clínicas com base na especialidade e UF fornecidas.",
        "parameters": [
          {
            "name": "especialidade",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "uf",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/api/clinicas/{id}": {
      "put": {
        "tags": [
          "Clínicas"
        ],
        "summary": "Atualiza uma clínica",
        "description": "Endpoint para atualizar uma clínica existente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Aceito"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      },
      "delete": {
        "tags": [
          "Clínicas"
        ],
        "summary": "Exclui uma clínica",
        "description": "Endpoint para excluir uma clínica pelo ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Não Encontrado"
          }
        }
      }
    },
    "/usuarios": {
      "post": {
        "tags": [
          "Usuários"
        ],
        "summary": "Efetua o logout  *",
        "description": "Endpoint para realizar logout",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      },
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Lista todos os usuários do sistema",
        "description": "Endpoint para obter todos os usuários do sistema.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/usuarios/login": {
      "post": {
        "tags": [
          "Usuários"
        ],
        "summary": "Efetua o login do usuário do sistema",
        "description": "Endpoint para efetuar o login e receber o token JWT.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Acesso Negado"
          },
          "404": {
            "description": "Não Encontrado"
          }
        }
      }
    },
    "/usuarios/id/{id}": {
      "get": {
        "tags": [
          "Usuários"
        ],
        "summary": "Exibe os detalhes do usuário logado",
        "description": "Endpoint para obter os detalhes do usuário logado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Não Encontrado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
    "/usuarios/id": {
      "put": {
        "tags": [
          "Usuários"
        ],
        "summary": "Permite edição dos dados do usuario",
        "description": "Endpoint para edita dados usuario.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Solicitação Inválida"
          },
          "404": {
            "description": "Não Encontrado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      },
      "delete": {
        "tags": [
          "Usuários"
        ],
        "summary": "Exclui um usuário",
        "description": "Endpoint para deletar usuario.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Não Encontrado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    }
  },
  "definitions": {
    "Erro": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "example": "Erro gerado pela aplicação"
        },
        "msg": {
          "type": "string",
          "example": "Mensagem detalhando o erro"
        },
        "param": {
          "type": "string",
          "example": "URL que gerou o erro"
        }
      }
    }
  }
}